+++ examples/example_config.py
@@ -1,5 +1,5 @@
 import naz
-from examples.example_klasses import ExampleRedisBroker, MySeqGen, MyRateLimiter
+from examples.example_klasses import ExampleRedisBroker  # , MySeqGen, MyRateLimiter
 
 
 # run as:
@@ -10,11 +10,7 @@ client = naz.Client(
     system_id="smppclient1",
     password="password",
     broker=ExampleRedisBroker(),
-    sequence_generator=MySeqGen(),
-    logger=naz.log.SimpleLogger(
-        "naz.client", level="INFO", log_metadata={"environment": "staging", "release": "canary"}
-    ),
-    enquire_link_interval=70.00,
-    rateLimiter=MyRateLimiter(),
-    address_range="^254",  # any msisdns beginning with 254. See Appendix A of SMPP spec doc
+    codec=naz.codec.SimpleCodec(encoding="ucs2"),
+    socket_timeout=4.00,
+    logger=naz.log.SimpleLogger("example_config.client", render_as_json=False),
 )