+++ naz/__init__.py
@@ -2,7 +2,7 @@ from .client import Client  # noqa: F401
 
 from . import log  # noqa: F401
 from . import broker  # noqa: F401
-from . import nazcodec  # noqa: F401
+from . import codec  # noqa: F401
 from . import protocol  # noqa: F401
 from . import throttle  # noqa: F401
 from . import sequence  # noqa: F401