+++ naz/protocol.py
@@ -3,7 +3,7 @@ from __future__ import annotations
 import json
 import typing
 
-from . import nazcodec
+from . import codec as the_codec
 
 
 class Message:
@@ -33,7 +33,7 @@ class Message:
         smpp_command: str,
         log_id: str,
         pdu: typing.Union[None, bytes] = None,
-        codec_class: typing.Union[None, nazcodec.BaseNazCodec] = None,
+        codec: typing.Union[None, the_codec.BaseCodec] = None,
         short_message: typing.Union[None, str] = None,
         source_addr: typing.Union[None, str] = None,
         destination_addr: typing.Union[None, str] = None,
@@ -45,7 +45,7 @@ class Message:
             smpp_command: any one of the SMSC commands eg submit_sm
             log_id: a unique identify of this reque
             pdu: the full PDU as sent to SMSC. It is mutually exclusive with `short_message`.
-            codec_class: python class instance to be used to encode/decode messages. It should be a child class of `naz.nazcodec.BaseNazCodec`.
+            codec: python class instance to be used to encode/decode messages. It should be a child class of `naz.codec.BaseCodec`.
                          You should only specify this, if you also specified `pdu`, else you can leave it as None.
             short_message: message to send to SMSC. It is mutually exclusive with `pdu`
             source_addr: the identifier(eg msisdn) of the message sender.
@@ -57,7 +57,7 @@ class Message:
             smpp_command=smpp_command,
             log_id=log_id,
             pdu=pdu,
-            codec_class=codec_class,
+            codec=codec,
             short_message=short_message,
             source_addr=source_addr,
             destination_addr=destination_addr,
@@ -67,7 +67,7 @@ class Message:
         self.smpp_command = smpp_command
         self.log_id = log_id
         self.pdu = pdu
-        self.codec_class = codec_class
+        self.codec = codec
         self.short_message = short_message
         self.source_addr = source_addr
         self.destination_addr = destination_addr
@@ -79,7 +79,7 @@ class Message:
         smpp_command: str,
         log_id: str,
         pdu: typing.Union[None, bytes],
-        codec_class: typing.Union[None, nazcodec.BaseNazCodec],
+        codec: typing.Union[None, the_codec.BaseCodec],
         short_message: typing.Union[None, str],
         source_addr: typing.Union[None, str],
         destination_addr: typing.Union[None, str],
@@ -129,16 +129,16 @@ class Message:
                     type(hook_metadata)
                 )
             )
-        if not isinstance(codec_class, (type(None), nazcodec.BaseNazCodec)):
+        if not isinstance(codec, (type(None), the_codec.BaseCodec)):
             raise ValueError(
-                "`codec_class` should be of type:: `None` or `naz.nazcodec.BaseNazCodec` You entered: {0}".format(
-                    type(codec_class)
+                "`codec` should be of type:: `None` or `naz.codec.BaseCodec` You entered: {0}".format(
+                    type(codec)
                 )
             )
-        if pdu and not codec_class:
+        if pdu and not codec:
             # because pdu is in bytes, when converting to string; we need to use whatever encoding was passed in
-            # thus one has to supply the codec_class
-            raise ValueError("You cannot specify `pdu` and not a `codec_class`")
+            # thus one has to supply the codec
+            raise ValueError("You cannot specify `pdu` and not a `codec`")
 
     def to_json(self) -> str:
         """
@@ -158,14 +158,14 @@ class Message:
         if self.pdu:
             if typing.TYPE_CHECKING:
                 # make mypy happy; https://github.com/python/mypy/issues/4805
-                assert isinstance(self.codec_class, nazcodec.BaseNazCodec)
-            _item["pdu"] = self.codec_class.decode(self.pdu)
+                assert isinstance(self.codec, the_codec.BaseCodec)
+            _item["pdu"] = self.codec.decode(self.pdu)
 
         return json.dumps(_item)
 
     @staticmethod
     def from_json(
-        json_message: str, codec_class: typing.Union[None, nazcodec.BaseNazCodec] = None
+        json_message: str, codec: typing.Union[None, the_codec.BaseCodec] = None
     ) -> Message:
         """
         Deserializes the message protocol from json. You can use this method if you would
@@ -173,21 +173,21 @@ class Message:
 
         Parameters:
             json_message: `naz.protocol.Message` in json format.
-            codec_class: python class instance to be used to encode/decode messages. It should be a child class of `naz.nazcodec.BaseNazCodec`.
+            codec: python class instance to be used to encode/decode messages. It should be a child class of `naz.codec.BaseCodec`.
                          You should only specify this, if `json_message` has a key called `pdu` and it is not None.
         """
         _in_dict = json.loads(json_message)
         if _in_dict.get("pdu"):
-            if not codec_class:
+            if not codec:
                 # because pdu is in bytes, when converting to string; we need to use whatever encoding was passed in
-                # thus one has to supply the codec_class
-                raise ValueError("You cannot specify `pdu` and not a `codec_class`")
+                # thus one has to supply the codec
+                raise ValueError("You cannot specify `pdu` and not a `codec`")
 
             # we need to convert pdu to bytes.
-            # There's a risk of a message been encoded with an codec_class that it was not decoded with when been saved to broker
-            # ideally, the codec_class should also be saved/serialized as part of the message been sent to broker.
+            # There's a risk of a message been encoded with an codec that it was not decoded with when been saved to broker
+            # ideally, the codec should also be saved/serialized as part of the message been sent to broker.
             # However, this risk is small and we should only consider it if people report it as a bug.
-            _in_dict["pdu"] = codec_class.encode(_in_dict["pdu"])
-            _in_dict["codec_class"] = codec_class
+            _in_dict["pdu"] = codec.encode(_in_dict["pdu"])
+            _in_dict["codec"] = codec
 
         return Message(**_in_dict)